refactor: move renamed class to correct file
